backend/requirements.txt
asgiref==3.10.0
Django==5.2.1
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-filter==25.2
django-unfold==0.72.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.4.0
pymongo==4.9.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
django-tailwind==4.4.1
pytailwindcss==0.3.0

